sort the utests + address comments
