+++ streams/src/test/java/org/apache/kafka/streams/state/internals/InMemoryWindowStoreTest.java
@@ -19,20 +19,36 @@ package org.apache.kafka.streams.state.internals;
 import org.apache.kafka.common.header.internals.RecordHeaders;
 import org.apache.kafka.common.serialization.Serde;
 import org.apache.kafka.common.serialization.Serdes;
+import org.apache.kafka.common.utils.Bytes;
 import org.apache.kafka.streams.KeyValue;
+import org.apache.kafka.streams.StreamsConfig;
 import org.apache.kafka.streams.kstream.Windowed;
 import org.apache.kafka.streams.processor.internals.ProcessorRecordContext;
 import org.apache.kafka.streams.query.Position;
+import org.apache.kafka.streams.query.PositionBound;
+import org.apache.kafka.streams.query.QueryConfig;
+import org.apache.kafka.streams.query.QueryResult;
+import org.apache.kafka.streams.query.WindowKeyQuery;
+import org.apache.kafka.streams.query.WindowRangeQuery;
 import org.apache.kafka.streams.state.KeyValueIterator;
 import org.apache.kafka.streams.state.StateSerdes;
 import org.apache.kafka.streams.state.Stores;
 import org.apache.kafka.streams.state.WindowStore;
 import org.apache.kafka.streams.state.WindowStoreIterator;
+import org.apache.kafka.test.InternalMockProcessorContext;
+import org.apache.kafka.test.StreamsTestUtils;
+import org.apache.kafka.test.TestUtils;
 
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
+import java.io.File;
+import java.time.Instant;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Properties;
 
 import static java.time.Duration.ofMillis;
 import static org.apache.kafka.common.utils.Utils.mkEntry;
@@ -40,6 +56,8 @@ import static org.apache.kafka.common.utils.Utils.mkMap;
 import static org.apache.kafka.streams.state.internals.WindowKeySchema.toStoreKeyBinary;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 
 public class InMemoryWindowStoreTest extends AbstractWindowBytesStoreTest {
@@ -185,4 +203,119 @@ public class InMemoryWindowStoreTest extends AbstractWindowBytesStoreTest {
         assertEquals(expected, actual);
     }
 
+    @Nested
+    class InMemoryStoreIQv2Tests {
+        private static final long WINDOW_SIZE = 10_000L;
+        private static final long RETENTION_PERIOD = 60_000L;
+        private static final long SEGMENT_INTERVAL = 30_000L;
+
+        private InMemoryWindowStore inMemoryStore;
+        private InternalMockProcessorContext<String, String> context;
+        private File baseDir;
+
+        @BeforeEach
+        public void setUp() {
+            final Properties props = StreamsTestUtils.getStreamsConfig();
+            baseDir = TestUtils.tempDirectory();
+            context = new InternalMockProcessorContext<>(
+                baseDir,
+                Serdes.String(),
+                Serdes.String(),
+                new StreamsConfig(props)
+            );
+
+            inMemoryStore = new InMemoryWindowStore(
+                "iqv2-inmemory-test-store",
+                RETENTION_PERIOD,
+                WINDOW_SIZE,
+                false,  // retainDuplicates
+                "test-metrics-scope"
+            );
+
+            inMemoryStore.init(context, inMemoryStore);
+        }
+
+        @AfterEach
+        public void tearDown() {
+            if (inMemoryStore != null) {
+                inMemoryStore.close();
+            }
+        }
+
+        @Test
+        public void shouldHandleWindowKeyQuerySuccessfullyOnInMemoryStore() {
+            final WindowKeyQuery<Bytes, byte[]> query = WindowKeyQuery.withKeyAndWindowStartRange(
+                new Bytes("test-key".getBytes()),
+                Instant.ofEpochMilli(0),
+                Instant.ofEpochMilli(Long.MAX_VALUE)
+            );
+            final PositionBound positionBound = PositionBound.unbounded();
+            final QueryConfig config = new QueryConfig(false);
+
+            final QueryResult<WindowStoreIterator<byte[]>> result = inMemoryStore.query(query, positionBound, config);
+
+            // Verify: In-memory stores support IQv2 queries
+            assertTrue(result.isSuccess(), "Expected query to succeed on in-memory store");
+            assertNotNull(result.getResult(), "Expected result iterator to be present");
+            assertNotNull(result.getPosition(), "Expected position to be set");
+        }
+
+        @Test
+        public void shouldHandleWindowRangeQuerySuccessfullyOnInMemoryStore() {
+            final WindowRangeQuery<Bytes, byte[]> query = WindowRangeQuery.withWindowStartRange(
+                Instant.ofEpochMilli(0),
+                Instant.ofEpochMilli(Long.MAX_VALUE)
+            );
+            final PositionBound positionBound = PositionBound.unbounded();
+            final QueryConfig config = new QueryConfig(false);
+
+            final QueryResult<KeyValueIterator<Windowed<Bytes>, byte[]>> result =
+                inMemoryStore.query(query, positionBound, config);
+
+            // Verify: In-memory stores support IQv2 queries
+            assertTrue(result.isSuccess(), "Expected query to succeed on in-memory store");
+            assertNotNull(result.getResult(), "Expected result iterator to be present");
+            assertNotNull(result.getPosition(), "Expected position to be set");
+        }
+
+        @Test
+        public void shouldCollectExecutionInfoForInMemoryStoreWhenRequested() {
+            final WindowKeyQuery<Bytes, byte[]> query = WindowKeyQuery.withKeyAndWindowStartRange(
+                new Bytes("test-key".getBytes()),
+                Instant.ofEpochMilli(0),
+                Instant.ofEpochMilli(Long.MAX_VALUE)
+            );
+            final PositionBound positionBound = PositionBound.unbounded();
+            final QueryConfig config = new QueryConfig(true); // Enable execution info
+
+            final QueryResult<WindowStoreIterator<byte[]>> result = inMemoryStore.query(query, positionBound, config);
+
+            // Verify: Execution info was collected
+            assertFalse(result.getExecutionInfo().isEmpty(), "Expected execution info to be collected");
+            boolean foundInMemoryStoreInfo = false;
+            for (final String info : result.getExecutionInfo()) {
+                if (info.contains("Handled in") && info.contains(InMemoryWindowStore.class.getName())) {
+                    foundInMemoryStoreInfo = true;
+                    break;
+                }
+            }
+            assertTrue(foundInMemoryStoreInfo, "Expected execution info to mention the in-memory store class");
+        }
+
+        @Test
+        public void shouldNotCollectExecutionInfoForInMemoryStoreWhenNotRequested() {
+            final WindowKeyQuery<Bytes, byte[]> query = WindowKeyQuery.withKeyAndWindowStartRange(
+                new Bytes("test-key".getBytes()),
+                Instant.ofEpochMilli(0),
+                Instant.ofEpochMilli(Long.MAX_VALUE)
+            );
+            final PositionBound positionBound = PositionBound.unbounded();
+            final QueryConfig config = new QueryConfig(false); // Disable execution info
+
+            final QueryResult<WindowStoreIterator<byte[]>> result = inMemoryStore.query(query, positionBound, config);
+
+            // Verify: No execution info was collected
+            assertTrue(result.getExecutionInfo().isEmpty(), "Expected no execution info to be collected");
+        }
+    }
 }
